* Fix review comments * Add newer dependency in java pom * Rename package name in pom * Fix version for formatter & remove build file * Fix package name * Add log for E2E test
