--- .travis.yml
@@ -1,6 +1,0 @@
-language: java
-jdk:
-  - oraclejdk8
-cache:
-  directories:
-    - $HOME/.m2
