--- src/main/java/org/apache/logging/log4j/ext/kusto/Constants.java
@@ -1,15 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-package org.apache.logging.log4j.ext.kusto;
-
-public class Constants {
-    public static final String LOG4J2_ADX_APP_ID = "LOG4J2_ADX_APP_ID";
-    public static final String LOG4J2_ADX_APP_KEY = "LOG4J2_ADX_APP_KEY";
-    public static final String LOG4J2_ADX_TENANT_ID = "LOG4J2_ADX_TENANT_ID";
-    public static final String LOG4J2_ADX_INGEST_CLUSTER_URL = "LOG4J2_ADX_INGEST_CLUSTER_URL";
-    public static final String INGESTION_RETRIES = "INGESTION_RETRIES";
-
-    private Constants() {
-
-    }
-}
