+++ src/test/java/com/microsoft/azure/kusto/log4j/log4j2.component.properties
